+++ src/app/boot.ejs.ts
@@ -11,6 +11,7 @@ import AppComponent from "app/AppComponent";
 import { BLOG_PROVIDERS } from "app/website/blog/Blog";
 import { APP_DEV_PROVIDERS } from "app/app-dev/AppDev";
 
+
 let appInjectorRef:Injector;
 export const appInjector = ( injector?:Injector ):Injector => {
 	if ( injector ) {